+++ A3A/addons/gear/CfgVehicles.hpp
@@ -1,0 +1,3 @@
+class CfgVehicles {
+	    #include "headgear\CfgVeh_headgear.hpp"
+}
