+++ A3A/addons/gear/CfgWeapons.hpp
@@ -1,0 +1,3 @@
+class CfgWeapons {
+	    #include "headgear\CfgWea_headgear.hpp"
+}
